requirements.txt
boto3==1.35.82
botocore==1.35.82
colorama==0.4.6
Faker==33.1.0
ffmpeg==1.4
matplotlib==3.10.0
numpy==2.2.0
openai==1.57.4
pandas==2.2.3
paramiko==3.5.0
Pillow==11.0.0
plexapi==4.16.0
pyclip==0.7.0
pyheif==0.8.0
pymongo==4.10.1
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.5
Requests==2.32.3
requests_html==0.10.0
spotipy==2.24.0
SQLAlchemy==2.0.36
streamlit==1.41.1
thefuzz==0.22.1
toml==0.10.2
tqdm==4.67.1
yt_dlp==2024.12.13
